+++ examples/web-wagmi/src/profiles/UseProfileFollowers.tsx
@@ -2,14 +2,14 @@ import { ProfileFragment } from '@lens-protocol/react';
 import { useState } from 'react';
 
 import { ProfileFollowers } from './components/ProfileFollowers';
-import { SelectProfile } from './components/ProfileSelector';
+import { ProfileSelector } from './components/ProfileSelector';
 
 export function UseProfileFollowers() {
   const [profile, setProfile] = useState<ProfileFragment | null>(null);
   return (
     <>
       <p>Select a profile:</p>
-      <SelectProfile onProfileSelected={(p) => setProfile(p)} />
+      <ProfileSelector onProfileSelected={(p) => setProfile(p)} />
       {profile && <ProfileFollowers profileId={profile.id} />}
     </>
   );