+++ examples/web-wagmi/src/publications/UsePublications.tsx
@@ -1,17 +1,22 @@
 import { usePublications } from '@lens-protocol/react';
 
+import { ErrorMessage } from '../components/error/ErrorMessage';
+import { Loading } from '../components/loading/Loading';
 import { useInfiniteScroll } from '../hooks/useInfiniteScroll';
 import { PublicationCard } from './components/PublicationCard';
 
 export function UsePublications() {
   const {
     data: publications,
+    error,
     loading,
     hasMore,
     observeRef,
   } = useInfiniteScroll(usePublications({ profileId: '0x15' }));
 
-  if (loading) return <div>Loading...</div>;
+  if (loading) return <Loading />;
+
+  if (error) return <ErrorMessage error={error} />;
 
   return (
     <div>