+++ packages/shared-kernel/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @lens-protocol/shared-kernel
 
+## 0.3.0
+
 ## 0.2.0
 
 ## 0.1.1